% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function 3.R
\name{bootstrap_CI}
\alias{bootstrap_CI}
\title{Bootstrapping}
\usage{
bootstrap_CI(x, y, alpha, bootstrap)
}
\arguments{
\item{x}{a \code{matrix} of predictor variables.}

\item{y}{a \code{vector} of binary response variables.}

\item{alpha}{a \code{numeric} of a significance level}

\item{bootstrap}{a \code{numeric} of the number of bootstraps to be performed}
}
\description{
Generate confidence intervals using a bootstrap method for
the logistic regression coefficients.
}
\examples{
x <-
y <-
alpha <- 0.05
bootstrap <- 20
bootstrap_CI(x, y, alpha, bootstrap)
}
\author{
Ava White

Bukola Ayodele
}
