% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function 2.R
\name{leastsquares}
\alias{leastsquares}
\title{Least Squares}
\usage{
leastsquares(x, y)
}
\arguments{
\item{x}{A \code{matrix} of predictor variables.}

\item{y}{A \code{vector} containing binary response variables (0 or 1).}
}
\value{
A \code{numeric} giving the value of \code{initial_values}
}
\description{
Generate initial values for optimization using the 'logreg'
function using the least squares formula.
}
\examples{
#generate random data
set.seed(123)
n <- 200
p <- 2
x <- matrix(rnorm(n * p), nrow = n, ncol = p)
y <- rbinom(n, 1, 0.5)
#use leastsquares function to generate initial values
initial_values <- leastsquares(x, y)
print(initial_values)
}
\author{
Ava White

Bukola Ayodele
}
