% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function 1.R
\name{logreg}
\alias{logreg}
\title{Logistic Regression}
\usage{
logreg(x, y)
}
\arguments{
\item{x}{A \code{matrix} of predictor variables.}

\item{y}{A \code{vector} containing binary response variables (0 or 1).}
}
\value{
A \code{numeric} giving the value of \code{beta_hat}
}
\description{
Perform a logistic regression using numerical optimization
to estimate the coefficient vector \code(beta) by minimizing the log-likelihood.
}
\examples{
#generate random data
set.seed(123)
n <- 100
p <- 2
x <- matrix(rnorm(n * p), nrow = p)
y <- rbinom(200)
#use logreg function to predict beta
beta_hat <- logreg(x, y)
}
\author{
Ava White

Bukola Ayodele
}
