% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function 5.R
\name{matrix_metrics}
\alias{matrix_metrics}
\title{Creating Confusion Matrix}
\usage{
matrix_metrics(y, pi_prob, cutoff = 0.5)
}
\arguments{
\item{y}{A \code{vector} of binary response variables.}

\item{pi_prob}{A \code{vector} of predicted probabilities.}

\item{cutoff}{A cut-off value for prediction that assigns '1' to values
higher and '0' to values lower.}
}
\value{
A confusion matrix and values for the 6 target metrics.
}
\description{
Create a table for visualizing how well a model performs,
specifically in terms of prevalence, accuracy, sensitivity, specificity,
false discovery rate, and diagnostic odds ratio.
}
\examples{
set.seed(123)
n <- 200
x <- cbind(1, rnorm(n))
beta_true <- c(-1, 2)
pi_true <- 1 / (1 + exp(-x \%*\% beta_true))
y <- rbinom(n, 1, pi_true)
result <- logreg(x, y)
pi_prob <- 1 / (1 + exp(-x \%*\% result))
metrics_result <- matrix_metrics(y, pi_prob)
print(metrics_result$confusion_matrix)
print(metrics_result$metrics)
}
\author{
Group 4
}
