% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function 4.R
\name{plot_logistic_curve}
\alias{plot_logistic_curve}
\title{Plotting a Logistic Curve}
\usage{
plot_logistic_curve(x, beta_hat, y)
}
\arguments{
\item{x}{A \code{matrix} of predictor variables.}

\item{beta_hat}{A \code{vector} of regression coefficients.}

\item{y}{A \code{vector} of binary response variables.}
}
\value{
A plot of a fitted logistic curve.
}
\description{
Plot a logistic curve that fits to the responses from a logistic
regression, where the x-axis represents a sequence of predicted beta values.
}
\examples{
set.seed(123)
n <- 200
x <- cbind(1, rnorm(n))
beta_true <- c(-1, 2)
pi_true <- 1 / (1 + exp(-x \%*\% beta_true))
y <- rbinom(n, 1, pi_true)
result <- logreg(x, y)
plot_logistic_curve(x, result, y)
}
\author{
Group 4
}
