% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function 6.R
\name{plot_metrics}
\alias{plot_metrics}
\title{Plotting Confusion Matrix Metrics}
\usage{
plot_metrics(y, pi_prob, metric, cutoff_grid = seq(0.1, 0.9, by = 0.1))
}
\arguments{
\item{y}{A \code{vector} of binary response variables.}

\item{pi_prob}{A \code{vector} of predicted probabilities.}

\item{metric}{Specifies which metric you want to visualize.}

\item{cutoff_grid}{A range of cutoff points.}
}
\value{
A plot of your chosen metric over a cutoff grid.
}
\description{
Create a plot that depicts the metrics of model performance
produced by the 'matrix_metrics' function (prevalence, accuracy,
sensitivity, specificity, false discovery rate, and diagnostic odds ratio)
}
\examples{
set.seed(123)
n <- 200
x <- cbind(1, rnorm(n))
beta_true <- c(-1, 2)
pi_true <- 1 / (1 + exp(-x \%*\% beta_true))
y <- rbinom(n, 1, pi_true)
plot_metrics(y, pi_prob, metric = "sensitivity")
plot_metrics(y, pi_prob, metric = "false_discovery_rate")
}
\author{
Group 4
}
