% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function 3.R
\name{bootstrap_CI}
\alias{bootstrap_CI}
\title{Bootstrapping}
\usage{
bootstrap_CI(x, y, alpha, bootstrap = 20)
}
\arguments{
\item{x}{A \code{matrix} of predictor variables.}

\item{y}{A \code{vector} of binary response variables.}

\item{alpha}{The significance level}

\item{bootstrap}{The number of bootstraps to be performed}
}
\value{
A \code{list} giving the values of \code{bootstrap_result} and
\code{CI_result}
}
\description{
Complete a bootstrap with a defined number of bootstraps and
significance level (alpha) and generate confidence intervals.
}
\examples{
#generate random data
set.seed(123)
n <- 200
p <- 2
x <- matrix(rnorm(n * p), nrow = n, ncol = p)
y <- rbinom(n, 1, 0.5)
alpha <- 0.05
bootstrap <- 20
bootstrap_result <- bootstrap_CI(x, y, alpha, bootstrap)
print(bootstrap_result)
}
\author{
Group 4
}
