% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function 1.R
\name{logreg}
\alias{logreg}
\title{Logistic Regression}
\usage{
logreg(x, y)
}
\arguments{
\item{x}{A \code{matrix} of predictor variables.}

\item{y}{A \code{vector} containing binary response variables (0 or 1).}
}
\value{
A \code{double} giving the value of \code{beta_hat}
}
\description{
Perform a logistic regression using numerical optimization
to estimate the coefficient vector \code{beta_hat} by minimizing the log-likelihood.
}
\examples{
#generate random data
set.seed(123)
n <- 200
p <- 2
x <- matrix(rnorm(n * p), nrow = n, ncol = p)
y <- rbinom(n, 1, 0.5)
#use logreg function to predict beta
beta_hat <- logreg(x, y)
print(beta_hat)
}
\author{
Group 4
}
